// Frontend/src/components/UI/Minimal/types/user.ts
export type IClassesItem = {
  id: number | undefined;
  member_id: number;
  payments_id: number;
  status: number;
	classes_date?: Date | string | null
  created_date?: Date;
	member?:IMemberItem;
	teacher?:IMemberItem;
	instrument?:IInstrumentItem;
	curriculum?:ICurriculumItem;
};
// ...
export type ICurriculumItem = {
  id?: number;
  instrument_id?: number;
  months: number;
  weekly_classes: number;
  total_classes: number;
  name: string;
  price: number;
  instrument_price: number;
  instrument_rental_fee: number;
  instrument_discount: number;
  image_file?: string;
  additional_text?: string;
  is_deleted: string;
  created_date: Date;
	instrument?:IInstrumentItem;
}





// 악기
export type IInstrumentItem = {
  id?: number;
  name: string;
  image_file1?: string;
  image_file2?: string;
  display_order: number;
  is_deleted: string;
  created_date: Date;
}





// 실시간 대화
export type ILiveChatItem = {
  id?: number;
  member_id: number;
  payments_id: number;
  content: string;
  created_date?: Date | null;
};




// 회원
export type IMemberItem = {
  id?: number;
  type: number;
  user_id: string;
  password: string;
  login_type?: string;
  login_ci_data?: string;
  name: string;
  nickname?: string;
  cellphone: string;
  image_file?: string;
  address1?: string;
  address2?: string;
  address3?: string;
  zip?: string;
  foreign_address?: string;
  gender: string;
  school_name?: string;
  grade?: string;
  memo?: string;
  zoom_link_url?: string;
  voov_link_url?: string;
  voov_link_exposed_members?: string;
  parent_name?: string;
  parent_cellphone?: string;
  total_payments?: number;
  total_price?: number;
  remaining_classes?: number;
  status: number;
  approve_date?: Date;
  leave_date?: Date | undefined;
  leave_message?: string;
  agree_marketing: string;
  notification_type: string;
  registration_source: string;
  last_login?: Date | undefined;
  created_date: Date;
};